+++ src/types.d.ts
@@ -145,7 +145,7 @@ declare global {
 
     interface ConsumableItem {
         system: {
-            charges: {
+            uses: {
                 value: number;
                 max: number;
             }